move DivoomAuraBoxFraming into own file
