+++ divoom_protocol_test.py
@@ -2,24 +2,6 @@ import unittest
 import divoom_protocol
 
 
-class TestDivoomAuraBoxFraming(unittest.TestCase):
-
-	def setUp(self):
-		self.testee = divoom_protocol.DivoomAuraBoxFraming()
-		pass
-
-	def test_replace_escaping_bytes(self):
-		# no replacements
-		self.assertEqual(self.testee.create([0, 7]), [1, 0, 7, 2])
-		self.assertEqual(self.testee.create([0, 7, 7]), [1, 0, 7, 7, 2])
-
-		# replace
-		self.assertEqual(self.testee.create([0, 1, 7]), [1, 0, 3, 4, 7, 2])
-		self.assertEqual(self.testee.create([2, 3, 7]), [1, 3, 5, 3, 6, 7, 2])
-		self.assertEqual(self.testee.create([2, 7]), [1, 3, 5, 7, 2])
-		self.assertEqual(self.testee.create([3]), [1, 3, 6, 2])
-
-
 class TestDivoomAuraBoxProtocol(unittest.TestCase):
 
 	TESTDATA_DIR = "testdata/"
@@ -213,6 +195,4 @@ class TestDivoomAuraBoxProtocol(unittest.TestCase):
 		# check with strings
 		test_package = self.testee.create_set_time_package("11", "30", "59")
 		self.assertEqual(test_package, package_assert)
-		
-if __name__ == '__main__':
-	unittest.main()
+
